ccgcn/requirements.txt
numpy
ordered-set==4.1.0
pandas==2.2.2
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
scikit-learn==1.5.0
scipy==1.13.0
setuptools==75.6.0
six==1.16.0
threadpoolctl==3.5.0
https://download.pytorch.org/whl/cu117/torch-1.13.1%2Bcu117-cp39-cp39-win_amd64.whl#sha256=e775fa85f412bd1bf816b8798dadb3b852b71e33e8008e9db29b6190ed94fe27
https://download.pytorch.org/whl/cu117/torchvision-0.14.1%2Bcu117-cp39-cp39-win_amd64.whl#sha256=c9fa805e58971ad0d8f4810b8e56f2b260939297ff564a988cf0194f31bd5d69
https://download.pytorch.org/whl/cu117/torchaudio-0.13.1%2Bcu117-cp39-cp39-win_amd64.whl#sha256=6b0a439212d5c57149526283b3e484fd9e99a418e683d91e8dc3d8d33ad0564c
torch_geometric==2.5.2
https://data.pyg.org/whl/torch-1.13.0%2Bcu117/torch_scatter-2.1.1%2Bpt113cu117-cp39-cp39-win_amd64.whl
tqdm==4.66.2
typing_extensions==4.9.0
wheel==0.45.1
yarl==1.9.3